refactor and reorganise repo
