+++ methods/test_sct_label_vertebrae.py
@@ -3,7 +3,7 @@ import numpy as np
 from spinalcordtoolbox.image import Image
 from spinalcordtoolbox.utils.sys import run_proc
 
-from utils import CONTRAST, edit_subject_lines_txt_file
+from utils.utils import CONTRAST, edit_subject_lines_txt_file
 
 
 #---------------------------Test Sct Label Vertebrae--------------------------
