+++ methods/test_spinenet_network.py
@@ -9,7 +9,7 @@ from spinalcordtoolbox.image import Image, get_dimension
 
 from spinenet import SpineNet, io
 
-from utils import CONTRAST, VERT_DISC, swap_y_origin, coord2list, project_on_spinal_cord, edit_subject_lines_txt_file 
+from utils.utils import CONTRAST, VERT_DISC, swap_y_origin, coord2list, project_on_spinal_cord, edit_subject_lines_txt_file 
 
 #---------------------------Test spinenet--------------------------
 def test_spinenet(args, test_mode=False):
